easygo_fastapi/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
transformers==4.37.2
torch==2.2.0
python-dateutil==2.8.2
pydantic==2.6.1
sentencepiece==0.1.99
openai==1.12.0
python-dotenv==1.0.1 